libs/pipeline_model/requirements.txt
grpcio>=1.0
